Major restructure performed
- Updated to anchor 0.30 and Solana program 1.18

- removed mpl-metadata crate and used anchor spl with metadata features

- updated freeze and thaw instructions

- create state and instructions folder
