+++ programs/anchor-nft-staking/Cargo.toml
@@ -14,8 +14,8 @@ no-idl = []
 no-log-ix-name = []
 cpi = ["no-entrypoint"]
 default = []
+idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
 
 [dependencies]
-anchor-lang = { version="0.25.0", features = ["init-if-needed"] }
-anchor-spl = "0.25.0"
-mpl-token-metadata = { version="1.4.1", features=["no-entrypoint"] }
+anchor-lang = { version="0.30.0", features = ["init-if-needed"] }
+anchor-spl = { version = "0.30.0", features = ["metadata"] }
