+++ programs/anchor-nft-staking/src/error.rs
@@ -1,0 +1,13 @@
+use anchor_lang::prelude::*;
+
+#[error_code]
+pub enum StakeError {
+    #[msg("NFT already staked")]
+    AlreadyStaked,
+
+    #[msg("State account is uninitialized")]
+    UninitializedAccount,
+
+    #[msg("Stake state is invalid")]
+    InvalidStakeState,
+}
