+++ programs/anchor-nft-staking/src/instructions/mod.rs
@@ -1,0 +1,7 @@
+pub mod stake;
+pub mod redeem;
+pub mod unstake;
+
+pub use stake::*;
+pub use redeem::*;
+pub use unstake::*;
