+++ programs/lootbox-program/Cargo.toml
@@ -14,8 +14,9 @@ no-idl = []
 no-log-ix-name = []
 cpi = ["no-entrypoint"]
 default = []
+idl-build = ["anchor-lang/idl-build", "anchor-spl/idl-build"]
 
 [dependencies]
-anchor-lang = { version="0.25.0", features=["init-if-needed"] }
-anchor-spl = "0.25.0"
+anchor-lang = { version="0.30.0", features=["init-if-needed"] }
+anchor-spl = "0.30.0"
 anchor-nft-staking = { path = "../anchor-nft-staking", features = ["cpi"] }
