+++ programs/lootbox-program/src/error.rs
@@ -1,0 +1,10 @@
+use anchor_lang::error_code;
+
+#[error_code]
+pub enum LootboxError {
+    #[msg("Mint already claimed")]
+    AlreadyClaimed,
+
+    #[msg("Haven't staked long enough for this loot box or invalid loot box number")]
+    InvalidLootbox,
+}
