+++ programs/lootbox-program/src/instructions/mod.rs
@@ -1,0 +1,5 @@
+pub mod open_box;
+pub mod retrieve_item;
+
+pub use open_box::*;
+pub use retrieve_item::*;
