+++ programs/lootbox-program/src/state/loot_box_pointer.rs
@@ -1,0 +1,9 @@
+use anchor_lang::prelude::*;
+
+#[account]
+#[derive(InitSpace)]
+pub struct LootboxPointer {
+    pub mint: Pubkey,
+    pub claimed: bool,
+    pub is_initialized: bool,
+}
