+++ programs/lootbox-program/src/state/mod.rs
@@ -1,0 +1,3 @@
+pub mod loot_box_pointer;
+
+pub use loot_box_pointer::*;
